requirements.txt
altgraph==0.17.4
auto-py-to-exe==2.44.2
bcrypt==4.2.0
bottle==0.13.2
bottle-websocket==0.2.9
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==43.0.3
Eel==0.17.0
ephem==4.1.5
future==1.0.0
gevent==24.10.3
gevent-websocket==0.10.1
gitdb==4.0.11
GitPython==3.1.43
greenlet==3.1.1
idna==3.10
macholib==1.16.3
packaging==24.1
paramiko==3.5.0
pillow==10.4.0
pycparser==2.22
pyinstaller==6.11.0
pyinstaller-hooks-contrib==2024.9
PyNaCl==1.5.0
pyparsing==3.2.0
requests==2.32.3
setuptools==75.3.0
smmap==5.0.1
urllib3==2.2.3
whichcraft==0.6.1
zope.event==5.0
zope.interface==7.1.1
